+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # HPO Workbench Changelog
 
+## v1.2.3
+* Added function to export HTML MONDO analysis file to system clipboard
 ## v.1.2.1
 * Extending HPO stats command to show number of subclasses and number of annotated diseases
+## v1.2.0
+* Improving disease search and new annotation suggestion
+* Fixing bug in Excel export
 
 ## v.1.1.0
 * Refactoring to use HpoDisease from phenol instead of old (and now removed) DiseaseModel class
-
-
 
 ## v.1.0.3
 * Refactoring to use phenol library 1.2.1